Extra cast to avoid warning; in this case, the truncation is intentional.
[git-p4: depot-paths = "//depot/openwatcom/": change = 35007]

+++ bld/sdk/wr/c/wrdmsg.c
@@ -24,8 +24,7 @@
 *
 *  ========================================================================
 *
-* Description:  WHEN YOU FIGURE OUT WHAT THIS FILE DOES, PLEASE
-*               DESCRIBE IT HERE!
+* Description:  Keep track of dialog messages.
 *
 ****************************************************************************/
 
@@ -88,7 +87,7 @@ int WR_EXPORT WRIsWRDialogMsg( MSG *msg )
     LIST        *item;
 
     for( item = WRDialogList; item != NULL ; item = ListNext( item ) ) {
-        hdlg = (HWND)ListElement( item );
+        hdlg = (HWND)(DWORD)ListElement( item );
         if( ( hdlg != (HWND)NULL ) && IsDialogMessage( hdlg, msg ) ) {
             return( TRUE );
         }
